Factor out Ferry into its own .hh and .cc files
